refactor: strengthen layering boundaries and review follow-ups
